third_party/blink/web_tests/fast/events/mouseenter-mouseleave-on-drag-expected.txt

Verifies that dragging mouse in/out of an element doesn't fire redundant mouseenter or mouseleave events (crbug.com/356090 & crbug.com/470258)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

--- drag across inner ---
outer received mouseover
outer received mouseenter
PASS wasInside['outer'] is false
outer received mousedown
outer received mouseout
inner received mouseover
inner received mouseenter
PASS wasInside['inner'] is false
inner received mouseout
inner received mouseleave
PASS wasInside['inner'] is true
outer received mouseover
outer received mouseup
outer received mouseout
outer received mouseleave
PASS wasInside['outer'] is true


--- drag into inner ---
outer received mouseover
outer received mouseenter
PASS wasInside['outer'] is false
outer received mousedown
outer received mouseout
inner received mouseover
inner received mouseenter
PASS wasInside['inner'] is false
inner received mouseup


--- drag out of inner ---
inner received mousedown
inner received mouseout
inner received mouseleave
PASS wasInside['inner'] is true
outer received mouseover
outer received mouseup


PASS successfullyParsed is true

TEST COMPLETE
